refactor: diffing times belongs to Event structure
